Geo locator package added

+++ lib/model/weather_model.dart
@@ -55,7 +55,7 @@ class Weather {
 }
 
 class Main {
-  int? temp;
+  double? temp;
   double? feelsLike;
   double? tempMin;
   double? tempMax;
